Owen feature branch (#48)
* Updated DataStorageAPIImpl and added ProductSumsCalculatorAPIImpl

added reference to EngineProcessAPI and took any methods that are related to calulcating the product-sum and moved them into a new file called ProductSumsCalculatorAPIImpl

* Updated ProductSumsCalculatorAPIImpl, ComputeOutput, and EngineOutput

Updated each of these files to use Lists rather than arrays.

* Added Constructors to DataStorageProcessAPIImpl

This ensures that the mock EngineProcessAPI is correctly put into DataStorageProcessAPIImpl, letting us control (and verify) the behavior on the tests

* Updated DataStorageAPI test package names

* Added DataStorageAPI smoke tests and integration tests

* Updated SmokeTest class name

* Updated ProductSumsCalculatorAPIImpl comments

* fixed OwenFeatureBranch inconsistencies

* Updated imports to not include *

checkstyle checks do not allow import *

* Updated comments for clarification

These files should be good to resolve Issue #42

* Commented out testLargeRange

Leaving this out of the tests for now while we think of a quicker more efficient solution. Was taking 2+ minutes on average

* Added test for processing data when there is no input/output paths

This resolves issue #43

* Updated DataStorageProcessAPIImpl to include 3 validation checks

1) validate the request object
2) validate the k range
3) validate the file paths

* Added unit test for validation check

Checking that the test file path validation works. this checks off step number 4 in assignment 5.